requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
edge-tts>=7.2.7,<8.0.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
